Resolved the failing tests:
a) there is a bug in apache camel that means that POST parameters are not parsed as expected, i.e. they are not parsed at all, leading to all the tests that use that failing.
b) the ApiConnector was appending an extra '&' to the end of the POST parameters, this is bad as IIS (What CCP uses) doesn't like it.